+++ httpx_auth/__init__.py
@@ -4,7 +4,7 @@ from httpx_auth._authentication import (
     QueryApiKey,
     SupportMultiAuth,
 )
-from httpx_auth._oauth2.common import OAuth2
+from httpx_auth._oauth2.common import OAuth2, DisplaySettings
 from httpx_auth._oauth2.authorization_code import (
     OAuth2AuthorizationCode,
     OktaAuthorizationCode,
@@ -47,6 +47,7 @@ __all__ = [
     "HeaderApiKey",
     "QueryApiKey",
     "OAuth2",
+    "DisplaySettings",
     "OAuth2AuthorizationCodePKCE",
     "OktaAuthorizationCodePKCE",
     "OAuth2Implicit",